+++ packages/core/loglayer/src/index.ts
@@ -10,15 +10,15 @@ export type {
   ErrorOnlyOpts,
   ILogBuilder,
   ILogLayer,
+  LogLayerContext,
+  LogLayerData,
+  LogLayerMetadata,
   LogLayerPlugin,
   LogLayerTransport,
   LogLevelType,
   PluginBeforeDataOutParams,
   PluginBeforeMessageOutParams,
   PluginShouldSendToLoggerParams,
-  LogLayerContext,
-  LogLayerMetadata,
-  LogLayerData
 } from "@loglayer/shared";
 export { LogLevel } from "@loglayer/shared";
 export { LogLayer } from "./LogLayer.js";